+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Run command for (without commit):
 - change list items spacing in `CheckboxListComponent` 
 - style for page container
 - loader component `<ng-loader>`
+- show loading indicator on Search page when recipes are loading
 
 ### 1.19.1